Add support for reading base64-encoded trust bundle from config map (#44)
* Add support for reading base64-encoded trust bundle from config map

The base64 trust bundle configmap can be provided via both Data as well
as BinaryData. The one in ConfigMap.Data is used as-is. The one in
ConfigMap.BinaryData is decoded using base64.

* Move certs in unit test to fixture in testdata

allows us to keep test fixtures isolated from test code.

* Update go.mod

Add dependency for reading yaml.

* Generate cert for testing

Remove testdata fixtures and generate certs for tests on the fly.